rename utils files
