+++ src/BookSearch/BookSearch.scss
@@ -594,372 +594,6 @@ form {
 	}
 }
 
-
-// .rating-form-wrapper {
-// 	display: flex;
-// 	background: linear-gradient(rgba(107, 209, 251, 0.95) 10%, rgba(107, 209, 251, 0.01));
-// 	flex-direction: column;
-// 	height: 0;
-// 	overflow: hidden;
-// 	position: relative;
-// 	top: 0;
-// 	bottom: 0;
-// 	z-index: 2;
-// 	form {
-// 		display: flex;
-// 		flex-direction: column;
-// 		align-items: center;
-
-// 		input:not[type="submit"], textarea {
-// 			width: 80%;
-// 			margin: -2px;
-// 			height: 30px;
-// 			border-radius: 4px;
-// 			padding: 2px;
-// 			font-size: 1.5rem;
-// 		}
-
-// 		label {
-// 			display: none;
-// 		}
-// 	}
-// }
-
-
-// // ----------------------- Ratings form -----------------------------
-
-// .form-show {
-// 	height: 100%;
-// 	max-height: 100%;
-// 	overflow: show;
-// 	max-width: 50vw;
-// 	/*height: 120%;*/
-// 	top: -192px;
-// 	transition: top 150ms ease, height 150ms ease;
-// }
-
-// figcaption:hover div.rating-title {
-// 	display: flex;
-// 	flex-direction: inline-flex;
-// }
-
-// .average-rating {
-// 	display: block;
-// 	height: 30px;
-// 	border: 2px solid transparent;
-// 	background: var(--main);
-// 	border-radius: 25px;
-// 	padding: 10px 15px;
-// 	width: 260px;
-// 	float: left;
-// 	clear: both;
-// }
-
-// .wide-rating {
-// 	width: 600px;
-// 	margin: 10px auto;
-// }
-
-// .average-rating span:first-child {
-// 	font-weight: bold;
-// 	background: transparent;
-// 	color: black;
-// }
-
-// .average-rating-text {
-// 	margin: 0; 
-// 	padding: 0;
-// 	font-size: 1.1rem;
-// 	// display: inline;
-// 	float: left;
-// }
-
-// .black-text {
-// 	color: black;
-// 	font-size: 1.3rem;
-// }
-
-// .bold {
-// 	font-weight: bold;
-// }
-
-// .average-rating-number {
-// 	float: right;
-// }
-
-// .book-details-and-ratings {
-// 	// background: rgba(107, 209, 251, 0.4);
-// 	position: static;
-// 	display: flex;
-// 	flex-direction: column;
-// 	justify-content: center;
-// 	align-items: center;
-// 	width: 50vw;
-// 	margin-top: 40px auto;
-
-// 	ul:first-child {
-// 		margin: 5px auto;
-// 		text-align: center;
-// 	}
-
-// 	ul {
-// 		display: flex;
-// 		flex-direction: column;
-// 		margin: 10px auto;
-// 		padding: 0;
-// 		align-items: center;
-// 	}
-// }
-
-// .name {
-// 	font-family: var(--font2);
-// 	margin: 0;
-// 	color: var(--linkfg);
-// 	display: inline;
-// }
-
-// .opaque-background {
-// 	background: var(--opaque);
-// }
-
-// .rating-number {
-// 	color: black;
-// 	font-weight: bold;	
-// 	padding: 2px 4px;
-// 	border-radius: 4px;
-// 	width: 50px;
-// 	margin: 0 auto;
-// 	text-align: center;
-// }
-
-// .more {
-// 	background: var(--gradient2);
-// }
-
-// // Sets rating colours on a scale from red to green
-// .rating-0 {
-// 	background: #FF0000;
-// }
-// .rating-1 {
-// 	background: #FF3300;
-// }
-// .rating-2 {
-// 	background: #ff6600;
-// }
-// .rating-3{
-// 	background: #ff9900;
-// }
-// .rating-4 {
-// 	background: #FFCC00;
-// }
-// .rating-5 {
-// 	background: #FFFF00;
-// }
-// .rating-6 {
-// 	background: #ccff00;
-// }
-// .rating-7 {
-// 	background: #99ff00;
-// }
-// .rating-8 {
-// 	background: #66ff00;
-// }
-// .rating-9 {
-// 	background: #33ff00;
-// }
-// .rating-10 {
-// 	background: #00FF00;
-// 	box-shadow: 3px -1px 6px 6px rgba(153,181,247,1);;
-// }	
-
-// .rating-notes {
-// 	display: none;
-// 	font-style: italic;
-// 	position: absolute;
-// 	background: rgba(107, 209, 251, 0.94);
-// 	margin-top: 40px;
-// 	width: 25vw;
-// 	padding: 5px;
-// 	left: 100px;
-// 	z-index: 10;
-// }
-
-// .rating {
-// 	position: relative;
-// }
-
-// .rating:hover .rating-notes {
-// 	display: block;
-// }
-
-// .book-image img {
-// 	width: 30vw;
-// 	height: auto;
-// 	min-height: 300px;
-// }
-
-// .book-title-wrapper, .book-author-wrapper {
-// 	// max-width: 500px;
-// }
-
-// .book-title {
-// 	background: var(--gradient2);
-// 	color: black;
-// 	padding: 0 1px;
-// }
-
-// .by {
-// 	font-size: 0.9rem;
-// 	background: var(--opaque);
-// 	// background: transparent;
-// 	color: var(--linkfg);
-// }
-
-// .book-author {
-// 	color: yellow;
-// 	background: var(--opaque);
-// 	font-size: 1.0rem;
-// 	margin: 0;
-// }
-
-// section {
-// 	display: flex;
-// 	min-height: 500px;
-// 	justify-content: center;
-// 	ul {
-// 		display: flex;
-// 		flex-direction: column;
-// 		margin: 10px 5%;
-// 		width: 100%;
-// 		text-align: left;
-// 		justify-content: center;
-// 		li {
-// 			margin: 10px;
-// 			span {
-// 				background: green;
-// 				color: black;
-// 				// border-radius: 4px;
-// 				padding: 0 2px;
-// 			}
-// 		}
-// 	}
-// }
-
-// .book-title-and-author {
-// 	width: 100%;
-// 	margin: 10px auto 20px;
-// }
-
-// .range-output {
-// 	width: 80px;
-// 	display: flex;
-// 	flex-direction: column;
-// 	margin: 5px auto;
-// }
-
-// // this is used with the range input for the ratings 
-// output { 
-// 	position: relative;
-// 	color: yellow;
-// 	font-size: 1.4rem;
-// 	background: black;
-// 	padding: 5px;
-// 	border-radius: 25px;
-// 	text-align: center; 
-// }
-
-// // // -------------------------- RANGE INPUT MIXINS --------------------------
-
-// // // Codepen to test this out: https://codepen.io/prupke/pen/wNVvJX
-
-
-// // @mixin range-inputs() { 
-// // 	background: skyblue; 
-// // 	font-size: 13rem;
-// // 	margin: 5px;
-// // 	padding: 0;
-// // 	width: 150px;
-// // 	height: 30px;
-// // }
-
-// // // set the width and height of the range input here
-// // $input-w: 12rem;
-// // $input-h: 2.0rem;
-// // $track-h: .625*$input-h;
-// // $thumb-d: .8125*$input-h;
-// // $filll-h: .3125*$input-h;
-
-// // // Adjust this if the tracker is misaligned (happens with webkit browsers)
-// // $sw: 0*$track-h;
-
-// // @mixin range-input-highlight-boxes($c: white, $a: .9) {
-// // 	box-sizing: border-box;
-// // 	// border: 2px solid skyblue;
-// // 	padding: none;
-// // 	border-radius: 25px;
-// // 	background: currentcolor content-box;
-// // 	color: rgba($c, $a);
-// // }
-
-// // @mixin range-input-track() {
-// // 	height: $track-h;
-// // 	@include range-input-highlight-boxes(white)
-// // }
-
-// // @mixin range-input-fill() {
-// // 	height: $filll-h;
-// // 	border: none;
-// // 	@include range-input-highlight-boxes(teal, .4)
-// // }
-
-// // @mixin range-input-thumb() {
-// // 	width: $thumb-d; height: $thumb-d;
-// // 	border-radius: 0;
-// // 	@include range-input-highlight-boxes(skyblue)
-// // }
-
-// // input[type='range'] {
-// // 	&, &::-webkit-slider-thumb {
-// // 		-webkit-appearance: none
-// // 	}
-// // 	border: none !important;
-// // 	margin: .25em auto;
-// // 	display: block;
-// // 	width: $input-w; height: $input-h;
-// // 	@include range-input-highlight-boxes;
-// // 	font-family: var(--font1);
-	
-// // 	&::-webkit-slider-runnable-track {
-// // 		@include range-input-track()
-// // 	}
-// // 	&::-moz-range-track {
-// // 		@include range-input-track()
-// // 	}
-// // 	&::-ms-track {
-// // 		@include range-input-track()
-// // 	}
-	
-// // 	&::-moz-range-progress {
-// // 		@include range-input-fill()
-// // 	}
-	
-// // 	&::-ms-fill-lower {
-// // 		@include range-input-fill()
-// // 	}
-	
-// // 	&::-webkit-slider-thumb {
-// // 		margin-top: .5*($track-h - $thumb-d) - 2*$sw;
-// // 		@include range-input-thumb()
-// // 	}
-// // 	&::-moz-range-thumb {
-// // 		@include range-input-thumb()
-// // 	}
-// // 	&::-ms-thumb {
-// // 		margin-top: 0;
-// // 		@include range-input-thumb()
-// // 	}
-// // }
-
 @media screen and (max-width: 800px) {
 	#logo {
 		width: calc(100vw - 150px);
