+++ src/ParseMock/utils.js
@@ -4,7 +4,7 @@ export const getBooksMock = (_clubId) => mockBooks;
 
 export const getBookMock = objectId => mockBooks.filter(book => book.objectId === objectId) || {};
 
-export const getRatingsMock = async (clubId) => mockRatings;
+export const getRatingsMock = async (_clubId) => mockRatings;
 
 export const getRatingsForBookMock = async (bookId, _clubId) => {
 	return mockRatings.filter(rating => rating.bookId === bookId) || [];
